:truck: Refactor applications + add ingressProps
